app/monthly_payroll/routes.py
from flask import Blueprint, render_template, request, jsonify, redirect, flash
from app.models import MonthlyPayroll  # Import the necessary models
from app.utils import user_belongs_to_company
from flask_login import login_required
from app.monthly_payroll import monthly_payroll_bp

# Define your routes
@monthly_payroll_bp.route('/<int:company_id>/<int:user_id>', methods=['GET'])
@login_required
@user_belongs_to_company
def monthly_payroll(company_id, user_id):
    # Your logic to retrieve and display the monthly payroll data goes here
    # You can fetch data from the MonthlyPayroll model or as needed

    return render_template('monthly_payroll.html', company_id=company_id, user_id=user_id, monthly_payroll_data=monthly_payroll_data)

@monthly_payroll_bp.route('/<int:company_id>/<int:user_id>/generate', methods=['POST'])
@login_required
@user_belongs_to_company
def generate_monthly_payroll(company_id, user_id):
    # Your logic to generate the monthly payroll goes here
    # You can perform calculations, create payroll records, and update the database

    flash('Monthly payroll generated successfully', 'success')
    return redirect(url_for('monthly_payroll.view_monthly_payroll', company_id=company_id, user_id=user_id))

@monthly_payroll_bp.route('/<int:company_id>/<int:user_id>/download', methods=['GET'])
@login_required
@user_belongs_to_company
def download_monthly_payroll(company_id, user_id):
    # Your logic to generate and serve the monthly payroll report for download goes here
    # You can create a downloadable file and return it as a response

    # For example, you can return a CSV file for download
    # Make sure to set the appropriate response headers

    response = make_response(generate_csv_monthly_payroll())
    response.headers["Content-Disposition"] = "attachment; filename=monthly_payroll.csv"
    response.headers["Content-Type"] = "text/csv"

    return response

# Add more routes and logic as needed

# Helper function to generate CSV content for monthly payroll report
def generate_csv_monthly_payroll():
    # Your logic to generate the CSV content goes here
    # You can use libraries like pandas or csv to create the CSV content

    # For example, you can use pandas to create a DataFrame and convert it to CSV
    import pandas as pd
    data = {
        # Your data goes here
    }
    df = pd.DataFrame(data)

    # Create the CSV content
    csv_content = df.to_csv(index=False)

    return csv_content
